# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ResolveInput {
  /** Bunkr album or file URL */
  url: string;
}

export interface BunkrFile {
  name: string;
  url: string;
  /**
     * File size in bytes
     * @nullable
     */
  size?: number | null;
  /** File type (image, video, other) */
  type: string;
  /**
     * Thumbnail URL for images/videos
     * @nullable
     */
  thumbnailUrl?: string | null;
  /**
     * Direct CDN download URL
     * @nullable
     */
  cdnUrl?: string | null;
}

export interface ResolveResult {
  /** @nullable */
  albumName?: string | null;
  totalFiles?: number;
  files: BunkrFile[];
}

export interface JobInput {
  files: BunkrFile[];
  archiveName: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface Job {
  id: string;
  status: JobStatus;
  totalFiles: number;
  processedFiles: number;
  archiveName: string;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}