# lib/api-zod/src/generated/types/bunkrFile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BunkrFile {
  name: string;
  url: string;
  /**
     * File size in bytes
     * @nullable
     */
  size?: number | null;
  /** File type (image, video, other) */
  type: string;
  /**
     * Thumbnail URL for images/videos
     * @nullable
     */
  thumbnailUrl?: string | null;
  /**
     * Direct CDN download URL
     * @nullable
     */
  cdnUrl?: string | null;
}