# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface Job {
  id: string;
  status: JobStatus;
  totalFiles: number;
  processedFiles: number;
  archiveName: string;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}