# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BunkrFile } from './bunkrFile';

export interface JobInput {
  files: BunkrFile[];
  archiveName: string;
}