# lib/api-zod/src/generated/types/resolveInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResolveInput {
  /** Bunkr album or file URL */
  url: string;
}