# lib/api-zod/src/generated/types/resolveResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BunkrFile } from './bunkrFile';

export interface ResolveResult {
  /** @nullable */
  albumName?: string | null;
  totalFiles?: number;
  files: BunkrFile[];
}